minor bug fixes in incremental processing

+++ components/analytics-processors/org.wso2.carbon.analytics.spark.core/src/main/java/org/wso2/carbon/analytics/spark/core/rdd/AnalyticsRDD.java
@@ -148,7 +148,7 @@ public class AnalyticsRDD extends RDD<Row> implements Serializable {
 
         private String incID;
 
-        private long incMaxTS = Integer.MIN_VALUE;
+        private long incMaxTS = Long.MIN_VALUE;
 
         public RowRecordIteratorAdaptor(Iterator<Record> recordItr, int tenantId,
                                         boolean incEnable, String incID) {
